Starting over on DAG
I didn't fully understand the assignment so my solution was not correct. I am going to switch to Kotlin for the solution and use breath first search. I also commented code in the .java files so they don't interfere with the Kotlin class e.g. the node class was interfering